requirements.txt
joblib==1.5.3
numpy==2.4.0
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.3
matplotlib
seaborn
torchvision
torch
